+++ templates/staff/index.html
@@ -51,7 +51,6 @@
         <a href="{% url 'staff:redirect-list' %}" class="list-group-item list-group-item-action">Redirects</a>
         <a href="{% url 'staff:repo-list' %}" class="list-group-item list-group-item-action">Repos</a>
         <a href="{% url 'staff:report-list' %}" class="list-group-item list-group-item-action">Reports</a>
-        <a href="{% url 'staff:report-publish-request-list' %}" class="list-group-item list-group-item-action">Report Publish Requests</a>
         <a href="{% url 'staff:user-list' %}" class="list-group-item list-group-item-action">Users</a>
         <a href="{% url 'staff:workspace-list' %}" class="list-group-item list-group-item-action">Workspaces</a>
       </div>
